% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test_helper.R
\name{my_t_test_helper}
\alias{my_t_test_helper}
\title{t-Test Helper: Automated Group-vs-Control or Two-Group Comparison with Multiple Testing Correction}
\usage{
my_t_test_helper(
  formula,
  data,
  ref.group = NULL,
  paired = FALSE,
  ask = TRUE,
  correction = TRUE
)
}
\arguments{
\item{formula}{Model formula (e.g. value ~ group).}

\item{data}{Dataframe containing the variables.}

\item{ref.group}{(Optional) The name of the reference (control) group (as string) for multiple group comparisons.}

\item{paired}{Logical. Should a paired test be performed? Default: FALSE.}

\item{ask}{Logical. Should the user be asked whether to proceed? Default: TRUE.}

\item{correction}{Logical. Should p-value adjustment for multiple testing be applied? Default: TRUE.}
}
\value{
Dataframe with test results and significance stars.
}
\description{
Checks for normality (Shapiro-Wilk) and homogeneity of variance (Brown-Forsythe test),
recommends the appropriate t-test or Wilcoxon test and multiple testing correction,
compares all groups against a reference group (if >2), or compares two groups,
and returns a result table with adjusted p-values and significance stars.
}
\examples{
# Example dataset
df <- data.frame(
  foldchange = c(1.2, 1.3, 0.9, 1.1, 1.0,2.0, 2.1, 1.9, 2.2, 2.0,
  1.5, 1.4, 1.6, 1.7, 1.5),
  group = c(rep("ctrl", 5),
  rep("treat1", 5),
  rep("treat2", 5))
)

# Run helper
my_t_test_helper(foldchange ~ group, data = df,
                 ref.group = "ctrl", ask = FALSE, correction = TRUE)
}
